[REFACTOR]: Rename jaseci.svcs and jaseci_serv.svcs to *.svc
