+++ jaseci_core/jaseci/actions/standard/mail.py
@@ -1,6 +1,6 @@
 from jaseci.actions.live_actions import jaseci_action
-from jaseci.svcs.mail.mail_svc import mail_svc
-from jaseci.svcs.common_svc import MAIL_ERR_MSG
+from jaseci.svc.mail.mail_svc import mail_svc
+from jaseci.svc.common_svc import MAIL_ERR_MSG
 from jaseci.utils.utils import logger
