+++ jaseci_core/jaseci/api/jac_api.py
@@ -5,7 +5,7 @@ from jaseci.api.interface import interface
 import os
 import json
 
-from jaseci.svcs.meta_svc import meta_svc
+from jaseci.svc.meta_svc import meta_svc
 
 
 class jac_api:
