+++ jaseci_core/jaseci/jsctl/jsctl.py
@@ -8,7 +8,7 @@ import os
 import pickle
 import functools
 import json
-from jaseci.svcs.meta_svc import meta_svc
+from jaseci.svc.meta_svc import meta_svc
 import requests
 
 from jaseci.utils.utils import copy_func
