+++ jaseci_core/jaseci/svc/mail/mail_svc.py
@@ -3,8 +3,8 @@ import ssl
 from smtplib import SMTP, SMTP_SSL
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
-from jaseci.svcs.common_svc import common_svc
-from jaseci.svcs.service_state import ServiceState as SS
+from jaseci.svc.common_svc import common_svc
+from jaseci.svc.service_state import ServiceState as SS
 from jaseci.utils.utils import logger
 
 ################################################
