+++ jaseci_core/jaseci/svc/task/task_common.py
@@ -13,7 +13,7 @@ DEFAULT_MSG = "Skipping scheduled walker!"
 
 class queue(Task):
     def run(self, wlk, nd, args):
-        from jaseci.svcs.meta_svc import meta_svc
+        from jaseci.svc.meta_svc import meta_svc
 
         hook = meta_svc().hook()
 
@@ -30,7 +30,7 @@ class scheduled_walker(Task):
         return self.hook.get_obj_from_store(UUID(jid))
 
     def run(self, name, ctx, nd=None, snt=None, mst=None):
-        from jaseci.svcs.meta_svc import meta_svc
+        from jaseci.svc.meta_svc import meta_svc
 
         self.hook = meta_svc().hook()
 
@@ -189,7 +189,7 @@ class scheduled_sequence(Task):
             holder[0][self.json_escape.sub("_", req[params])] = holder[1]
 
     def trigger_interface(self, req: dict):
-        from jaseci.svcs.meta_svc import meta_svc
+        from jaseci.svc.meta_svc import meta_svc
 
         master = req.get("master")
         app = meta_svc()
