+++ jaseci_core/jaseci/svc/task/task_svc.py
@@ -1,10 +1,10 @@
 import signal
 import sys
 from multiprocessing import Process
-from jaseci.svcs.common_svc import common_svc
-from jaseci.svcs.service_state import ServiceState as SS
+from jaseci.svc.common_svc import common_svc
+from jaseci.svc.service_state import ServiceState as SS
 from jaseci.utils.utils import logger
-from jaseci.svcs.task.task_common import (
+from jaseci.svc.task.task_common import (
     task_properties,
     queue,
     scheduled_walker,
