+++ jaseci_core/jaseci/tests/test_jac.py
@@ -1,7 +1,7 @@
 from antlr4 import InputStream, CommonTokenStream
 from jaseci.jac.jac_parse.jacLexer import jacLexer
 from jaseci.jac.jac_parse.jacParser import jacParser
-from jaseci.svcs.meta_svc import meta_svc
+from jaseci.svc.meta_svc import meta_svc
 from jaseci.actor.sentinel import sentinel
 from jaseci.graph.graph import graph
 import jaseci.actions.live_actions as lact
