+++ jaseci_serv/jaseci_serv/base/models.py
@@ -12,7 +12,7 @@ from django.contrib.auth import get_user_model
 from jaseci.element.master import master as core_master
 from jaseci.element.super_master import super_master as core_super
 from jaseci.api.interface import interface
-from jaseci_serv.svcs.meta_svc import meta_svc
+from jaseci_serv.svc.meta_svc import meta_svc
 
 
 class master(core_master):
