+++ jaseci_serv/jaseci_serv/base/tests/test_models.py
@@ -5,7 +5,7 @@ from jaseci_serv.base import models
 from jaseci_serv.obj_api.views import JaseciObjectSerializer
 from jaseci.element import element
 from jaseci.element.super_master import super_master
-from jaseci_serv.svcs.meta_svc import meta_svc
+from jaseci_serv.svc.meta_svc import meta_svc
 
 
 def sample_user(email="JSCITEST_user@jaseci.com", password="whatever"):
