+++ jaseci_serv/jaseci_serv/svc/meta_svc.py
@@ -1,7 +1,7 @@
-from jaseci.svcs.meta_svc import meta_svc as ms
-from jaseci_serv.svcs.redis.redis_svc import redis_svc
-from jaseci_serv.svcs.task.task_svc import task_svc
-from jaseci_serv.svcs.mail.mail_svc import mail_svc
+from jaseci.svc.meta_svc import meta_svc as ms
+from jaseci_serv.svc.redis.redis_svc import redis_svc
+from jaseci_serv.svc.task.task_svc import task_svc
+from jaseci_serv.svc.mail.mail_svc import mail_svc
 from jaseci_serv.jaseci_serv.settings import RUN_SVCS
