+++ jaseci_serv/jaseci_serv/svc/redis/redis_svc.py
@@ -2,7 +2,7 @@
 This module includes code related to hooking Jaseci's Redis to the
 core engine.
 """
-from jaseci.svcs.redis.redis_svc import redis_svc as rs
+from jaseci.svc.redis.redis_svc import redis_svc as rs
 from jaseci_serv.jaseci_serv.settings import REDIS_CONFIG
