+++ jaseci_serv/jaseci_serv/svc/task/task_svc.py
@@ -1,5 +1,5 @@
 from jaseci_serv.jaseci_serv.settings import TASK_CONFIG
-from jaseci.svcs.task.task_svc import task_svc as ts
+from jaseci.svc.task.task_svc import task_svc as ts
 
 
 #################################################
